Add support for deletion of listens from BigQuery
Deletion works as follows:
* When the user deletes their account, a job is pushed into
the queue.
* The bigquery_job_runner then processes the queue and
deletes users from the BigQuery table.

This required some refactoring. I did the following stuff:
* moved stuff like code to run queries etc. into a `bigquery` module,
* moved the stats_calculator into a more generic bigquery_job_runner
