+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-testpaths = listenbrainz/db listenbrainz/webserver listenbrainz/listenstore listenbrainz/tests/unit listenbrainz/stats listenbrainz/bigquery_writer
+testpaths = listenbrainz/db listenbrainz/webserver listenbrainz/listenstore listenbrainz/tests/unit listenbrainz/stats listenbrainz/bigquery_writer listenbrainz/bigquery_job_runner
 addopts = --cov=listenbrainz
